Replaced the spring-boot:build-image (buildkit) with google jib. For some reason that thing keeps throwing the following error: [ERROR] Failed to execute goal org.springframework.boot:spring-boot-maven-plugin:2.3.0.RELEASE:build-image (default-cli) on project ci: Execution default-cli of goal org.springframework.boot:spring-boot-maven-plugin:2.3.0.RELEASE:build-image failed: Docker API call to 'localhost/v1.24/images/create?fromImage=gcr.io%2Fpaketo-buildpacks%2Fbuilder%3Abase-platform-api-0.3' failed with status code 500 "com.sun.jna.LastErrorException: [2] No such file or directory" -> [Help 1]
Have no patience to test further.

So chuck that.

Going for Jib. Dont have to write my own Dockerfile which is good. And it also uploads. So no need of the exec plugin thingy also. But not removing it since it can be kept for future reference (Running some other kind of command)

SO here goes first attempt!

Gode!